+++ src/palau/lims/setuphandlers.py
@@ -7,6 +7,7 @@
 import os
 
 from bika.lims import api
+from bika.lims.api import security
 from bika.lims.browser.analysisrequest.add2 import AR_CONFIGURATION_STORAGE
 from BTrees.OOBTree import OOBTree
 from palau.lims import logger
@@ -32,6 +33,7 @@ from senaite.ast.config import AST_CALCULATION_TITLE
 from senaite.ast.config import IDENTIFICATION_KEY
 from senaite.ast.config import SERVICE_CATEGORY
 from senaite.ast.config import SERVICES_SETTINGS
+from senaite.core import permissions as permissionsCore
 from senaite.core.catalog import ANALYSIS_CATALOG
 from senaite.core.catalog import SAMPLE_CATALOG
 from senaite.core.catalog import SETUP_CATALOG
@@ -89,7 +91,7 @@ SETUP_FOLDERS = [
 # Tuples of (portal_type, list of behaviors)
 BEHAVIORS = [
     ("SampleContainer", [
-        "palau.lims.behaviors.samplecontainer.IExtendedSampleContainerBehavior",
+        "palau.lims.behaviors.samplecontainer.IExtendedSampleContainerBehavior",  # noqa
     ]),
     ("Patient", [
         "palau.lims.behaviors.patient.IExtendedPatientBehavior"
@@ -178,9 +180,9 @@ WORKFLOWS_TO_UPDATE = {
                 "new_state": "",
                 "action": "Create supplementary test",
                 "guard": {
-                    "guard_permissions": permissions.TransitionCreateSupplementary,
+                    "guard_permissions": permissions.TransitionCreateSupplementary,  # noqa
                     "guard_roles": "",
-                    "guard_expr": "python:here.guard_handler('create_supplementary')",
+                    "guard_expr": "python:here.guard_handler('create_supplementary')",  # noqa
                 }
             }
         }
@@ -217,6 +219,22 @@ WORKFLOWS_TO_UPDATE = {
 }
 
 
+# List of tuples of (role, path, [(permissions, acquire), ...])
+# When path is empty/None, the permissions are applied to portal object
+ROLES = [
+    ("Rejector", "", [
+        (permissionsCore.TransitionRejectAnalysis, 0),
+    ]),
+]
+
+# List of tuples of (id, title, roles)
+GROUPS = [
+    ("Rejectors", "Rejectors", [
+        "Member", "Rejector", "Analyst"
+    ]),
+]
+
+
 def setup_handler(context):
     """Generic setup handler
     """
@@ -231,6 +249,9 @@ def setup_handler(context):
 
     # Setup catalogs
     setup_catalogs(portal)
+
+    # Setup roles and groups
+    setup_roles_and_groups(portal)
 
     # Setup folders
     add_setup_folders(portal)
@@ -499,7 +520,7 @@ def hide_action(folder, action_id):
                 return n
         return -1
 
-    logger.info("Hide {} from control_panel".format(action_id, item.Title()))
+    logger.info("Hide {} from control_panel".format(action_id, item.Title()))  # noqa
     cp = api.get_tool("portal_controlpanel")
     action_index = get_action_index(action_id)
     if action_index == -1:
@@ -697,7 +718,9 @@ def update_workflow_state(workflow, status_id, settings):
 
     # Set basic info (title, description, etc.)
     new_status.title = settings.get("title", new_status.title)
-    new_status.description = settings.get("description", new_status.description)
+    new_status.description = settings.get(
+        "description", new_status.description
+    )
 
     # Set transitions
     trans = settings.get("transitions", ())
@@ -870,3 +893,47 @@ def setup_languages(portal):
     for key, val in LANG_SETTINGS:
         ploneapi.portal.set_registry_record("plone.{}".format(key), val)
     logger.info("Setup languages settings [DONE]")
+
+
+def setup_roles(portal):
+    """Setup roles
+    """
+    logger.info("Setup roles ...")
+    for role, path, perms in ROLES:
+        folder_path = path or api.get_path(portal)
+        folder = api.get_object_by_path(folder_path)
+        for permission, acquire in perms:
+            logger.info("{} {} {} (acquire={})".format(role, folder_path,
+                                                       permission, acquire))
+            security.grant_permission_for(folder, permission, role,
+                                          acquire=acquire)
+    logger.info("Setup roles [DONE]")
+
+
+def setup_groups(portal):
+    """Setup roles and groups
+    """
+    logger.info("Setup groups ...")
+    portal_groups = api.get_tool("portal_groups")
+    for group_id, title, roles in GROUPS:
+
+        # create the group and grant the roles
+        if group_id not in portal_groups.listGroupIds():
+            logger.info("Adding group {} ({}): {}".format(
+                title, group_id, ", ".join(roles)))
+            portal_groups.addGroup(group_id, title=title, roles=roles)
+
+        # grant the roles to the existing group
+        else:
+            ploneapi.group.grant_roles(groupname=group_id, roles=roles)
+            logger.info("Granting roles for group {} ({}): {}".format(
+                title, group_id, ", ".join(roles)))
+
+    logger.info("Setup groups [DONE]")
+
+
+def setup_roles_and_groups(portal):
+    """Setup roles and groups
+    """
+    setup_roles(portal)
+    setup_groups(portal)